Fix owner/SLA CI: relocate sla into graph package, null owner in data layer
Three CI gates the local run missed:
- Package layout: sla.py was a new top-level src/agent_bom module (forbidden);
  move it to graph/sla.py, co-located with the severity policy it depends on.
- Export/API contract: an unassigned finding's owner is now an explicit None
  (honest absence preserved in JSON/SARIF/CDX/SPDX and /v1/findings) rather than
  the display string 'Unassigned'; the CLI and UI render None as 'Unassigned' at
  the presentation layer only. Restores the findings-facets export contract.
- Product metrics: refresh docs/PRODUCT_METRICS.json module count (840 -> 841).
