+++ docs/PRODUCT_METRICS.json
@@ -40,7 +40,7 @@
     },
     {
       "name": "Python modules",
-      "value": 840,
+      "value": 841,
       "source": "src/agent_bom",
       "notes": "Counts all Python files recursively."
     },
